[wp]优化启动脚本，加入minimist读取命令参数

+++ gulp/develop.js
@@ -3,6 +3,7 @@
 let gulp = require('gulp');
 let exec = require('child_process').exec;
 let spawn = require('child_process').spawn;
+var argv = require('minimist')(process.argv.slice(2));
 
 let sass = require('gulp-sass');
 
@@ -10,11 +11,14 @@ let sass = require('gulp-sass');
  * 开发环境
  */
 gulp.task('server', () => {
-    // 切换环境并启动服务
-    exec('jekyll serve').stdout.on('data', (chunk) => {
-        console.log(chunk);
-    });
 
+    var env = argv.env || 'develop';
+
+    // 切换环境并启动服务
+    switchEnv(env);
+    runServer();
+
+    // 监控scss变化
     gulp.watch('src/static/scss/**/*.scss', ['sass']);
 });
 
@@ -29,23 +33,23 @@ gulp.task('watch', () => {
     gulp.watch('src/static/scss/**/*.scss', ['sass']);
 });
 
-// 切换环境
-gulp.task('develop', () => {
-    switchEnv('develop');
-    gulp.run('server');
-});
-
-gulp.task('production', () => {
-    switchEnv('production');
-    gulp.run('server');
-});
-
 let switchEnv = (ENV) => {
+    // 若不为dev或prod，则不切换环境
+    if (ENV !== 'develop' && ENV !== 'production') {
+        return;
+    }
+
     exec('sed -i.tmp \'s/^ENV:.*$/ENV: ' + ENV + '/\' _config.yml && rm _config.yml.tmp', (err, stdout) => {
         if (err) {
             console.log('Switch env: ' + err);
         } else {
-            console.log('Switch env: ' + ENV );
+            console.log('Switch env: ' + ENV);
         }
+    });
+}
+
+let runServer = () => {
+    exec('jekyll serve').stdout.on('data', (chunk) => {
+        console.log(chunk);
     });
 }
